fix rest reader unit test with mock properties

+++ src/main/java/org/tctalent/anonymization/batch/BatchProperties.java
@@ -27,5 +27,5 @@ public class BatchProperties {
   private int chunkSize;
   private int pageSize;
   private int maxReadSkips;
-  private int fetchDelayMillis;
+  private long fetchDelayMillis;
 }
